core/requirements.txt
Django==5.0.6
django-mailer==2.3.1
mysqlclient==2.2.4
Pillow==10.3.0
